Split pure sidebar logic into a service
