+++ src/app/components/elements/layout/SidebarLayout.tsx
@@ -32,6 +32,7 @@ import { getProgressIcon } from "../../pages/Gameboard";
 import { tags as tagsService } from "../../../services";
 import { Markup } from "../markup";
 import { History } from "history";
+import { calculateSidebarLink, containsActiveTab, isSidebarGroup } from "../../../services/sidebar";
 
 export const SidebarLayout = (props: RowProps) => {
     const { className, ...rest } = props;
@@ -1477,41 +1478,6 @@ export const GlossarySidebar = (props: GlossarySidebarProps) => {
             </>}
         </search>
     </ContentSidebar>;
-};
-
-export const calculateSidebarLink = (entry: SidebarEntryDTO): string | undefined => {
-    switch (entry.pageType) {
-        case "isaacBookDetailPage": {
-            const detailPageSplit = entry.pageId?.split(BOOK_DETAIL_ID_SEPARATOR);
-            if (!detailPageSplit || detailPageSplit.length !== 2) {
-                return undefined;
-            }
-            return `/${documentTypePathPrefix[DOCUMENT_TYPE.BOOK_INDEX_PAGE]}/${detailPageSplit[0].slice("book_".length)}/${detailPageSplit[1]}`;
-        }
-        case "isaacBookIndexPage": {
-            return `/${documentTypePathPrefix[DOCUMENT_TYPE.BOOK_INDEX_PAGE]}/${entry.pageId?.slice(`book_`.length)}`;
-        }
-        case "isaacRevisionDetailPage": {
-            return `/${documentTypePathPrefix[DOCUMENT_TYPE.REVISION]}/${entry.pageId}`;
-        }
-        case "page": {
-            return `/${documentTypePathPrefix[DOCUMENT_TYPE.GENERIC]}/${entry.pageId}`;
-        }
-    }
-    return undefined;
-};
-
-const isSidebarGroup = (entry: SidebarEntryDTO): entry is SidebarGroupDTO => {
-    return entry.type === "sidebarGroup";
-};
-
-const containsActiveTab = (group: SidebarGroupDTO, currentPathname: string): boolean => {
-    return !!group.sidebarEntries?.some(subEntry => {
-        if (isSidebarGroup(subEntry)) {
-            return containsActiveTab(subEntry, currentPathname);
-        }
-        return currentPathname === calculateSidebarLink(subEntry);
-    });
 };
 
 const SidebarEntries = ({ entry, history }: { entry: SidebarEntryDTO, history: History }) => {
